# swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: event api service
schemes:
# tip: remove http to make production-grade
- http
- https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - multipart/form-data
  - text/plain
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/plain
  - text/csv
  - text/html
  - image/png
  - image/gif
  - image/jpeg
host: localhost
basePath: /
...
paths:
...
  /customers/{id}:
    x-swagger-router-controller: customer
    get:
      summary: Get Customer by id
      description: Get Customer by id
      operationId: getCustomer
      tags: 
        - Customer
      parameters:
        - in: header
          name: authorization
          description: 'jwt authorization token'
          required: false
          type: string
        - in: path
          name: id
          description: 'id of Customer to retrieve'
          required: true
          type: string
        - in: query
          name: _select
          description: 'select only specified properties eg &#39;username,email&#39; exclude &#39;-email,-password&#39;'
          required: false
          type: string
      responses:
        default:
          description: Default Response 
          schema:
            $ref: "#/definitions/DefaultResponse"
    put:
      summary: Update Customer by id
      description: Update Customer by id
      operationId: updateCustomer
      tags: 
        - Customer
      parameters:
        - in: header
          name: authorization
          description: 'jwt authorization token'
          required: false
          type: string
        - in: path
          name: id
          description: 'id of Customer to update'
          required: true
          type: string
        - in: body
          name: data
          schema:
            $ref: "#/definitions/UpdateCustomer"
      responses:
        default:
          description: Default Response 
          schema:
            $ref: "#/definitions/DefaultResponse"
    delete:
      summary: Delete Customer by id
      description: Delete Customer by id
      operationId: deleteCustomer
      tags: 
        - Customer
      parameters:
        - in: header
          name: authorization
          description: 'jwt authorization token'
          required: false
          type: string
        - in: path
          name: id
          description: 'id of Customer to delete'
          required: true
          type: string
      responses:
        default:
          description: Default Response 
          schema:
            $ref: "#/definitions/DefaultResponse"
  /customers:
    x-swagger-router-controller: customer
    get:
      summary: Get all Customers
      description: Get all Customers
      operationId: getCustomers
      tags: 
        - Customer
      parameters:
        - in: header
          name: authorization
          description: 'jwt authorization token'
          required: false
          type: string
        - in: query
          name: _limit
          description: 'limit response length'
          required: false
          type: number
        - in: query
          name: _skip
          description: 'skip the given amount of data'
          required: false
          type: number
        - in: query
          name: _sort
          description: 'sort by data eg &quot;createdAt:1&quot; in ASC, &quot;createdAt:-1&quot; in DESC'
          required: false
          type: string
        - in: query
          name: _count
          description: 'if specified, returns the number count of query result'
          required: false
          type: boolean
        - in: query
          name: _select
          description: 'select only specified properties eg &#39;username,email&#39; exclude &#39;-email,-password&#39;'
          required: false
          type: string
        - in: query
          name: firstName
          description: 'firstName property'
          required: false
          type: string
        - in: query
          name: lastName
          description: 'lastName property'
          required: false
          type: string
        - in: query
          name: email
          description: 'email property'
          required: false
          type: string
        - in: query
          name: phone
          description: 'phone property'
          required: false
          type: string
        - in: query
          name: createdAt
          description: 'date of creation'
          required: false
          type: string
        - in: query
          name: updatedAt
          description: 'date of last update'
          required: false
          type: string
      responses:
        default:
          description: Default Response 
          schema:
            $ref: "#/definitions/DefaultResponse"
    post:
      summary: Create new Customer
      description: Create new Customer
      operationId: createCustomer
      tags: 
        - Customer
      parameters:
        - in: header
          name: authorization
          description: 'jwt authorization token'
          required: false
          type: string
        - in: body
          name: data
          schema:
            $ref: "#/definitions/CreateCustomer"
      responses:
        default:
          description: Default Response 
          schema:
            $ref: "#/definitions/DefaultResponse"
...
definitions:
...
  CreateCustomer:
    type: object
    required:
      - firstName
      - email
      - phone
    properties:
      firstName:
        type: string
        description: 'firstName property'
      lastName:
        type: string
        description: 'lastName property'
      email:
        type: string
        description: 'email property'
      phone:
        type: string
        description: 'phone property'
  UpdateCustomer:
    type: object
    properties:
      firstName:
        type: string
        description: 'firstName property'
      lastName:
        type: string
        description: 'lastName property'
      email:
        type: string
        description: 'email property'
      phone:
        type: string
        description: 'phone property'
...
  DefaultResponse: {}